backend/requirements.txt
aioredis==2.0.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
dj-rest-auth==6.0.0
Django==5.0.6
django-cors-headers==4.3.1
django-otp==1.5.0
django-prometheus==2.3.1
django-redis==5.4.0
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.19
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.2
pytest==8.2.1
pytest-asyncio==0.23.7
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.1
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==12.0
zope.interface==6.4.post0
